requirements.txt
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-thumbs==0.4
django-thumbs-v2==0.4.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
